Make SSM parameter module bulletproof
Terraform modules using `count` are susceptible to the items in a list being shifted when any element is added or removed from the list. This takes place as a result of each element's index changing.

In the case of some resources like SSM variables, this can cause data loss. Any time an element in the middle of the list is inserted or removed, the index of each element below the removed element change; this results in the parameter being removed and then recreated. However, in the process of doing so, the parameter value (which is maintained externally in the case of sensitive data) is lost.

Module is refactored to define a map of maps instead of a list of maps. The "outer" map consists of keys representing parameter names; the values stored under each key consists of another map that may contain the keys description, type, and value.

By this technique, each SSM variable is referenced by its unique key (the variable name) rather than by an index. This should allow multiple SSM variables to be managed in a single configuration directory without risk of parameter values being lost as the result of unrelated additions or deletions to other parameters.